FGenerators.jl

FGenerators

FGenerators.@fgenerator

FGenerators.@yield

FGenerators — Module

FGenerators: foldl for humans™

FGenerators.jl is a package for defining Transducers.jl-compatible extended foldl with a simple @yield-based syntax. An example for creating an ad-hoc "generator":

julia> using FGenerators

julia> @fgenerator function generate123()
           @yield 1
           @yield 2
           @yield 3
       end;

julia> collect(generate123())
3-element Array{Int64,1}:
 1
 2
 3

julia> sum(generate123())
6

FGenerators.jl is a spin-off of GeneratorsX.jl.

Use FLoops.jl to iterate over the items yielded from the generator:

julia> using FLoops

julia> @floop for x in generate123()
           @show x
       end
x = 1
x = 2
x = 3

FGenerators.@fgenerator — Macro

@fgenerator function f(...) ... end
@fgenerator f(...) = ...
@fgenerator(generator::GeneratorType) do; ...; end

Define a function f that returns a "generator" usable with Transducers.jl-compatible APIs (akd foldable interface). In @fgenerator(generator::GeneratorType) do ... end form, define Transducers.jl interface for GeneratorType.

Use @yield in the function body for producing an item. Use return to finish producing items.

See also FGenerators.

Extended help

Examples

Defining a function that returns a generator:

julia> using FGenerators

julia> @fgenerator function generate123()
           @yield 1
           @yield 2
           @yield 3
       end;

julia> collect(generate123())
3-element Array{Int64,1}:
 1
 2
 3

Defining foldable interface for a pre-existing type:

julia> struct Counting end;

julia> @fgenerator(itr::Counting) do
           i = 1
           while true
               @yield i
               i += 1
           end
       end;

julia> using Transducers  # for `Take`

julia> Counting() |> Take(3) |> collect
3-element Array{Int64,1}:
 1
 2
 3

Note that function such as collect and sum still dispatches to iterate-based methods (above Counting example worked because Counting was wrapped by Take). To automatically dispatch to foldl-based methods, use Foldable as the supertype:

julia> struct Count <: Foldable
           start::Int
           stop::Int
       end;

julia> @fgenerator(itr::Count) do
           i = itr.start
           i > itr.stop && return
           while true
               @yield i
               i == itr.stop && break
               i += 1
           end
       end;

julia> collect(Count(0, 2))
3-element Array{Int64,1}:
 0
 1
 2

FGenerators.@yield — Macro

@yield item

Yield an item from a generator. This is usable only inside special contexts such as within @fgenerator macro.